modules/trend/requirements.txt
PyYAML
attrs
asyncpg
sqlalchemy[asyncio]
redis
python-dotenv
aiohttp
fastapi
uvicorn[standard]
